EXCURSIA
This "EXCURSIA" is the flutter application design of toursit hidden spots in kerala.This was just the designs of login , signup and home page
main.dart :- The design of Home page of "EXCURSIA". 
loginpage.dart :-The design of Login and Sign up page.
Kasragod.dart :-The design of Kasargod hidden spots.